coaches and owners can view participant list
changed calendar.dart so that it displays a list of "class_card" components instead of hard_coded boxes with class info. each component will show class info, coach, date/time, and a list of participants registered for the class. the list of participants will only show for coach and owner accounts.